requirements.txt
# Install this package
-e .
# Install additional dependencies

# Install helpful dependencies (python 3.8)
certifi==2023.11.17
flake8==7.0.0
ipykernel==6.29.0
ipython==8.12.3
jupyter==1.0.0
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib==3.7.4
matplotlib-inline==0.1.6
numpy==1.24.4
pandas==2.0.3
pillow==10.2.0
plotly==5.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.10.1
seaborn==0.13.2
urllib3==2.1.0